pyproject.toml
[tool.poetry]
name = "np.com.bsi.places"
version = "0.1.0"
description = "A backend for places app"
authors = ["Mithun Adhikari <mithunadhikari40@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.22.1"
djangorestframework-simplejwt = "^5.1.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
